type comparisons: fixes
This is a huge patch that should be a global NOP.

For unix type platforms it enables -Wconversion to issue warnings (-> error)
for all automatic casts that seem less than ideal but are normally concealed
by the toolchain.

This is things like passing an int to a size_t argument.  Once enabled, I
went through all args on my default build (which build most things) and
tried to make the removed default cast explicit.

With that approach it neither change nor bloat the code, since it compiles
to whatever it was doing before, just with the casts made explicit... in a
few cases I changed some length args from int to size_t but largely left
the causes alone.

From now on, new code that is relying on less than ideal casting
will complain and nudge me to improve it by warnings.

+++ lib/secure-streams/protocols/ss-h2.c
@@ -176,7 +176,7 @@ secstream_tx_credit_add_h2(lws_ss_handle_t *h, int add)
 {
 	lwsl_info("%s: h %p: add %d\n", __func__, h, add);
 	if (h->wsi)
-		return lws_h2_update_peer_txcredit(h->wsi, LWS_H2_STREAM_SID, add);
+		return lws_h2_update_peer_txcredit(h->wsi, (unsigned int)LWS_H2_STREAM_SID, add);
 
 	return 0;
 }
